Modified   docs/PushButtonArray.png Modified   include/Common.h Modified   include/Controls/Control.h Modified   include/Controls/Potentiometer.h Modified   include/Controls/PushButtonArray.h Modified   include/DSP/Test.h Modified   include/DSP/Wah.h Modified   include/Math.h Modified   include/Wave/BiquadFilter.h Modified   include/Wave/HighPassFilter.h Modified   include/Wave/LowPassFilter.h Modified   include/Wave/NoiseGate.h Modified   include/Wave/WaveShaper.h

+++ include/Wave/WaveShaper.h
@@ -11,7 +11,7 @@ class WaveShaper : public Wave
 public:
 	WaveShaper(void)
 	{
-		for (uint8 i = 0; i < TABLE_SIZE; ++i)
+		for (uint16 i = 0; i < TABLE_SIZE; ++i)
 		{
 			float x = ((float)i - 127.5f) / 127.5f;
